test(item_extraction): unify typed and untyped tests under same shape
