--- tests/specs/item_extraction/item_extraction_spec.lua
@@ -1,24 +1,0 @@
-package.path = package.path .. ";" .. debug.getinfo(1, "S").source:sub(2):match "(.*/)" .. "/tests"
-
-local test_utils = require "tests.utils"
-local IGNORE = {
-	"markdown",
-	"html",
-}
-local LANGS_TO_TEST = vim.iter(require("codedocs").get_supported_langs())
-	:filter(function(v) return not vim.list_contains(IGNORE, v) end)
-	:totable()
-
-describe("Basic item extraction", function()
-	for _, lang in ipairs(LANGS_TO_TEST) do
-		for _, cases in pairs(require("tests.specs.item_extraction.test_cases." .. lang)) do
-			for idx, case in ipairs(cases) do
-				it(lang .. " - Case #" .. idx, function()
-					test_utils.mock_buffer(lang, case.structure, { row = case.cursor_pos, col = 0 })
-
-					assert.are.same(case.expected_items, require("codedocs").extract_item_data(lang))
-				end)
-			end
-		end
-	end
-end)
