--- tests/specs/item_extraction/test_cases/c.lua
@@ -1,61 +1,0 @@
-return {
-	func = {
-		{
-			structure = {
-				"void foo() {",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {},
-			},
-		},
-		{
-			structure = {
-				"int foo() {",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {
-					{
-						name = "",
-						type = "int",
-					},
-				},
-			},
-		},
-		{
-			structure = {
-				"int add(int a, int b, int c) {",
-				"    return a + b",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {
-					{
-						name = "a",
-						type = "int",
-					},
-					{
-						name = "b",
-						type = "int",
-					},
-					{
-						name = "c",
-						type = "int",
-					},
-				},
-				returns = {
-					{
-						name = "",
-						type = "int",
-					},
-				},
-			},
-		},
-	},
-}
