--- tests/specs/item_extraction/test_cases/java.lua
@@ -1,84 +1,0 @@
-return {
-	func = {
-		{
-			structure = {
-				"public void foo() {",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {},
-			},
-		},
-		{
-			structure = {
-				"public int foo() {",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {
-					{
-						name = "",
-						type = "int",
-					},
-				},
-			},
-		},
-		{
-			structure = {
-				"public int foo(int a, int b, int c) {",
-				"    return a;",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {
-					{
-						name = "a",
-						type = "int",
-					},
-					{
-						name = "b",
-						type = "int",
-					},
-					{
-						name = "c",
-						type = "int",
-					},
-				},
-				returns = {
-					{
-						name = "",
-						type = "int",
-					},
-				},
-			},
-		},
-	},
-	class = {
-		{
-			structure = {
-				"public class Foo {",
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				attributes = {},
-			},
-		},
-		{
-			structure = {
-				"public class Foo {",
-				'	public String a = "";',
-				"}",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				attributes = {},
-			},
-		},
-	},
-}
