--- tests/specs/item_extraction/test_cases/python.lua
@@ -1,77 +1,0 @@
-return {
-	func = {
-		{
-			structure = {
-				"def foo():",
-				"	pass",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {},
-			},
-		},
-		{
-			structure = {
-				"def foo() -> bool:",
-				"	pass",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {},
-				returns = {
-					{
-						name = "",
-						type = "bool",
-					},
-				},
-			},
-		},
-		{
-			structure = {
-				"def foo(a, b, c):",
-				"	return a",
-			},
-			cursor_pos = 1,
-			expected_items = {
-				parameters = {
-					{
-						name = "a",
-						type = "",
-					},
-					{
-						name = "b",
-						type = "",
-					},
-					{
-						name = "c",
-						type = "",
-					},
-				},
-				returns = {
-					{
-						name = "",
-						type = "",
-					},
-				},
-			},
-		},
-	},
-	class = {
-		{
-			structure = {
-				"class Foo:",
-				'	a = ""',
-			},
-			cursor_pos = 1,
-			expected_items = {
-				attributes = {
-					{
-						name = "a",
-						type = "",
-					},
-				},
-			},
-		},
-	},
-}
